pyproject.toml
[tool.poetry]
name = "crewai-albeorla"
version = "0.1.0"
description = ""
authors = ["Albert Orlando <albertjorlando+git@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
gitpython = "^3.1.41"
pygithub = "^2.1.1"
google-api-python-client = "^2.113.0"
todoist-python = "^8.1.4"
notion-client = "^2.2.1"
crewai = "^0.1.32"
unstructured = '==0.10.25'
tools = "^0.1.9"
duckduckgo-search = "^4.2"
loguru = "^0.7.2"
langchain-community = "^0.0.13"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

